requirements.txt
Django==1.11.6
djangorestframework==3.7.1
pytz==2017.2
